[grid-template] Renamed file with utils for grid
