+++ lib/hacks/grid-area.js
@@ -1,5 +1,5 @@
 const Declaration = require('../declaration');
-const shorthand = require('./grid-shorthand');
+const utils = require('./grid-utils');
 
 class GridArea extends Declaration {
 
@@ -15,10 +15,10 @@ class GridArea extends Declaration {
             return undefined;
         }
 
-        const values = shorthand.parse(decl);
+        const values = utils.parse(decl);
 
-        const [rowStart, rowSpan] = shorthand.translate(values, 0, 2);
-        const [columnStart, columnSpan] = shorthand.translate(values, 1, 3);
+        const [rowStart, rowSpan] = utils.translate(values, 0, 2);
+        const [columnStart, columnSpan] = utils.translate(values, 1, 3);
 
         if (rowStart) {
             decl.cloneBefore({
