+++ lib/hacks/grid-row-column.js
@@ -1,5 +1,5 @@
 const Declaration = require('../declaration');
-const shorthand = require('./grid-shorthand');
+const utils = require('./grid-utils');
 
 class GridRowColumn extends Declaration {
 
@@ -11,8 +11,8 @@ class GridRowColumn extends Declaration {
     insert(decl, prefix, prefixes) {
         if (prefix !== '-ms-') return super.insert(decl, prefix, prefixes);
 
-        const values = shorthand.parse(decl);
-        const [start, span] = shorthand.translate(values, 0, 1);
+        const values = utils.parse(decl);
+        const [start, span] = utils.translate(values, 0, 1);
 
         if (start) {
             decl.cloneBefore({
