+++ lib/hacks/grid-rows-columns.js
@@ -1,5 +1,5 @@
 const Declaration = require('../declaration');
-const shorthand = require('./grid-shorthand');
+const utils = require('./grid-utils');
 
 class GridRowsColumns extends Declaration {
 
@@ -31,7 +31,7 @@ class GridRowsColumns extends Declaration {
      */
     set(decl, prefix) {
         if (prefix === '-ms-' && decl.value.indexOf('repeat(') !== -1) {
-            decl.value = shorthand.changeRepeat(decl.value);
+            decl.value = utils.changeRepeat(decl.value);
         }
         return super.set(decl, prefix);
     }
