+++ lib/hacks/grid-template.js
@@ -1,5 +1,5 @@
 const Declaration = require('../declaration');
-const shorthand = require('./grid-shorthand');
+const utils = require('./grid-utils');
 
 class GridTemplate extends Declaration {
 
@@ -30,19 +30,19 @@ class GridTemplate extends Declaration {
         const [
             templateRows,
             templateColumns
-        ] = shorthand.parseTemplateShortcut(decl);
+        ] = utils.parseTemplateShortcut(decl);
 
         if (templateRows) {
             decl.cloneBefore({
                 prop: '-ms-grid-rows',
-                value: shorthand.changeRepeat(templateRows.join(''))
+                value: utils.changeRepeat(templateRows.join(''))
             });
         }
 
         if (templateColumns) {
             decl.cloneBefore({
                 prop: '-ms-grid-columns',
-                value: shorthand.changeRepeat(templateColumns.join(''))
+                value: utils.changeRepeat(templateColumns.join(''))
             });
         }
